Add initial version of EventRecorder.
The suite allows you to record test cases for web applications, meaning that
all user interaction with a web page will be recorded. This includes URL loads,
touch events and hardware keyboard events. You can also evaluate JavaScript and
capture screenshots, and the result of these actions will be saved on the host
computer and can later be compared to a previous test run.

EventRecorder consists of a Python script running on a host that records all
touch and hardware keyboard events, as well as allowing the user to perform a
few actions described below.

In addition, there's a Java application running on your target Android device.
The device application uses an Android WebView component that is able to
record and replay all user interaction. This component is the exact same one
used in the Android browser.

The Python script communicates with this application via messages or file
sending over ADB. If you're familiar with our RemoteJS tool, you'll have a
better understanding on how it works already, since we used the same techniques
for this suite.